% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kt_get.R
\name{ktb_project_data}
\alias{ktb_project_data}
\title{Get data of project}
\usage{
ktb_project_data(endpoint, project_id, username, password)
}
\arguments{
\item{endpoint}{api endpoint url}

\item{project_id}{project id}

\item{username}{username}

\item{password}{password}
}
\value{
data.frame
}
\description{
Get data of project
}
\examples{
\dontrun{
ktb_project_data(endpoint = "https://kc.humanitarianresponse.info/api/v1/",
                 project_id = 1,
                 username = Sys.getenv("KTB_USER"),
                 password = Sys.getenv("KTB_PWD")
)
}
}
