% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kt_get.R
\name{ktb_projects_list}
\alias{ktb_projects_list}
\title{Get list of projects}
\usage{
ktb_projects_list(endpoint, username, password)
}
\arguments{
\item{endpoint}{api endpoint url}

\item{username}{username}

\item{password}{password}
}
\value{
data.frame
}
\description{
Get list of projects
}
\examples{
\dontrun{
ktb_projects_list(endpoint = "https://kc.humanitarianresponse.info/api/v1/",
                  username = Sys.getenv("KTB_USER"),
                  password = Sys.getenv("KTB_PWD"))
}
}
